A2A/pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "openai-agents>=0.3.0",
    "uvicorn>=0.35.0",
]
